+++ recycler-decorator/lib-recycler-decorator-ktx/src/main/java/ru/surfstudio/android/recycler/decorator/ktx/BaseItemControllerDecoration.kt
@@ -1,4 +1,4 @@
-package ru.surfstudio.android.recycler.decorator.sample.easydecor
+package ru.surfstudio.android.recycler.decorator.ktx
 
 import android.graphics.Canvas
 import android.view.View
@@ -8,6 +8,9 @@ import ru.surfstudio.android.easyadapter.item.BaseItem
 import ru.surfstudio.android.easyadapter.item.NoDataItem
 import ru.surfstudio.android.recycler.decorator.Decorator
 
+/**
+ * Wrapper for connect Decorator.ViewHolderDecor and EasyAdapter
+ */
 @Suppress("UNCHECKED_CAST")
 class BaseItemControllerDecoration<I : BaseItem<out RecyclerView.ViewHolder>>(
         private val baseViewHolderDecor: BaseViewHolderDecor<I>