+++ recycler-decorator/lib-recycler-decorator-ktx/src/main/java/ru/surfstudio/android/recycler/decorator/ktx/BaseViewHolderDecor.kt
@@ -1,10 +1,13 @@
-package ru.surfstudio.android.recycler.decorator.sample.easydecor
+package ru.surfstudio.android.recycler.decorator.ktx
 
 import android.graphics.Canvas
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import ru.surfstudio.android.easyadapter.item.BaseItem
 
+/**
+ * Base ViewHolderDecor for EasyAdapter
+ */
 interface BaseViewHolderDecor<I: BaseItem<out RecyclerView.ViewHolder>> {
 
     fun draw(canvas: Canvas,