+++ recycler-decorator/lib-recycler-decorator-ktx/src/main/java/ru/surfstudio/android/recycler/decorator/ktx/BaseViewHolderOffset.kt
@@ -1,10 +1,13 @@
-package ru.surfstudio.android.recycler.decorator.sample.easydecor
+package ru.surfstudio.android.recycler.decorator.ktx
 
 import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import ru.surfstudio.android.easyadapter.item.BaseItem
 
+/**
+ * Base ViewHolderOffset decor for EasyAdapter
+ */
 interface BaseViewHolderOffset<I : BaseItem<out RecyclerView.ViewHolder>> {
     fun getItemOffsets(outRect: Rect,
                        view: View,