+++ recycler-decorator/sample/src/main/java/ru/surfstudio/android/recycler/decorator/sample/easyadapter/chat/ChatActivity.kt
@@ -14,9 +14,9 @@ import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.controll
 import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.decor.ChatDecorOffset
 import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.decor.ChatMessageDecor
 import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.decor.MessageTimeDecor
-import ru.surfstudio.android.recycler.decorator.sample.easydecor.offset
-import ru.surfstudio.android.recycler.decorator.sample.easydecor.overlay
-import ru.surfstudio.android.recycler.decorator.sample.easydecor.underlay
+import ru.surfstudio.android.recycler.decorator.ktx.offset
+import ru.surfstudio.android.recycler.decorator.ktx.overlay
+import ru.surfstudio.android.recycler.decorator.ktx.underlay
 
 class ChatActivity : AppCompatActivity() {