+++ recycler-decorator/sample/src/main/java/ru/surfstudio/android/recycler/decorator/sample/easyadapter/chat/decor/MessageTimeDecor.kt
@@ -2,22 +2,14 @@ package ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.decor
 
 import android.content.Context
 import android.graphics.*
-import android.util.Log
 import android.view.View
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.synthetic.main.chat_message_layout.view.*
 import kotlinx.android.synthetic.main.item_controller_message_time.view.*
-import ru.surfstudio.android.easyadapter.EasyAdapter
 import ru.surfstudio.android.easyadapter.item.BindableItem
-import ru.surfstudio.android.logger.Logger
-import ru.surfstudio.android.recycler.decorator.Decorator
 import ru.surfstudio.android.recycler.decorator.sample.R
-import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.ChatMessageDirection
-import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.ChatObject
-import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.controller.ChatMessageController
 import ru.surfstudio.android.recycler.decorator.sample.easyadapter.chat.controller.MessageTimeController
-import ru.surfstudio.android.recycler.decorator.sample.easydecor.BaseViewHolderDecor
+import ru.surfstudio.android.recycler.decorator.ktx.BaseViewHolderDecor
 import ru.surfstudio.android.recycler.decorator.sample.toPx
 
 class MessageTimeDecor(private val context: Context) : BaseViewHolderDecor<BindableItem<String, MessageTimeController.Holder>> {