+++ recycler-decorator/RELEASE_NOTES.md
@@ -4,3 +4,4 @@
 ## 0.5.0-alpha
 ##### Recycler-decorator
 * ANDDEP-785 added Builder for RecyclerView.ItemDecoration and sample
+* ANDDEP-785 added ktx for Decorator.Builder for use with EasyAdapter